Move all lua files to main dir
